+++ 1.05.04-wordcount-ctype.c
@@ -10,7 +10,7 @@ int main() {
 
 	bool within_word = false;
 	wint_t c;
-	size_t nl, nw, nc;
+	long nl, nw, nc;
 
 	nl = nw = nc = 0;
 	while ((c = getwchar()) != WEOF) {
@@ -28,6 +28,6 @@ int main() {
 		}
 	}
 
-	wprintf(L"%zu %zu %zu\n", nl, nw, nc);
+	wprintf(L"%ld %ld %ld\n", nl, nw, nc);
 	return 0;
 }
